FEAT: Quote script
Added a script that should in theory grab a quote from any message it was replied to. If it was a picture, then it'll grab a URL of that picture instead.
It adds a unique ID to each quote.
a random quote function can be called, or the user can fetch a quote by the ID.
moderators and admins can delete quotes based on ID